fastapi_pro_project/requirements.txt

fastapi
uvicorn
httpx
python-jose
pydantic
redis
